Move NewsLoaderPresentationAdapter to separate file
